+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Generalise input and output parameters of createRedirectedProcess
+* Breaking change: Generalise input and output parameters of createRedirectedProcess. Existing programs should replace `Process t` with `Process t ByteString ByteString` and `ProcessConfig t` with `ProcessConfig t ByteString`.
 
 ## 0.1.0.1