+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Revision history for reflex-process
 
-## Unreleased
+## 0.2.0.0
 
 * Breaking change: Generalise input and output parameters of createRedirectedProcess. Existing programs should replace `Process t` with `Process t ByteString ByteString` and `ProcessConfig t` with `ProcessConfig t ByteString`.